requirements.txt
absl-py
astor
audioread
certifi
cycler
decorator
gast
grpcio
h5py
joblib
Keras
Keras-Applications
Keras-Preprocessing
kiwisolver
librosa
llvmlite
Markdown
matplotlib
numba
numpy
#object-detection
pandas
protobuf
pyparsing
python-dateutil
python-speech-features
pytz
PyYAML
resampy
scikit-learn
scipy
six
tensorboard
tensorflow-gpu
termcolor
tqdm
Werkzeug
wincertstore
